requirements.txt
pdfplumber>=0.11
python-docx>=1.1
